Move DeviceProperties object to core, implement API payload formatting as extension functions and adjust how we handled test fixtures
