+++ sdk/analytics/src/main/java/com/klaviyo/analytics/networking/requests/UnregisterPushTokenApiRequest.kt
@@ -1,7 +1,7 @@
 package com.klaviyo.analytics.networking.requests
 
-import com.klaviyo.analytics.DeviceProperties
 import com.klaviyo.analytics.model.Profile
+import com.klaviyo.core.DeviceProperties
 import com.klaviyo.core.Registry
 
 /**
